Made pant a double instead of decimal as it just makes more sense structure wise.
Pant wont have its own seperate BLL as that seems redudant unless we decide that Pant needs to be pulled statistics on specifically.

+++ DataTransferObject/Model/LiquidProductDto.cs
@@ -6,7 +6,7 @@ public abstract class LiquidProductDto : ProductDto
     public int VolumeCl { get; set; }
     public decimal SalesPrice { get; set; }
 
-    public decimal Pant { get; set; }
+    public double Pant { get; set; }
 
 
     protected LiquidProductDto(string name, decimal costPrice, int stockQuantity,int volumeCl, decimal salesPrice) : base(name, costPrice, stockQuantity)
@@ -15,7 +15,7 @@ public abstract class LiquidProductDto : ProductDto
         SalesPrice = salesPrice;
     }
 
-    protected LiquidProductDto(string name, decimal Pant, decimal costPrice, int stockQuantity, int volumeCl, decimal salesPrice) : base(name, costPrice, stockQuantity)
+    protected LiquidProductDto(string name, double Pant, decimal costPrice, int stockQuantity, int volumeCl, decimal salesPrice) : base(name, costPrice, stockQuantity)
     {
         VolumeCl = volumeCl;
         SalesPrice = salesPrice;
